[#7] rename : AdminMemberService 폴더 이동
